working on search
